refactor: 전체 리스트 화면에서 각각의 아이템에 대한 imgUrl
