+++ app/src/main/java/com/sdhong/pokemonapp/viewmodel/AllPokemonViewModel.kt
@@ -3,10 +3,13 @@ package com.sdhong.pokemonapp.viewmodel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.sdhong.pokemonapp.model.Pokemon
-import com.sdhong.pokemonapp.model.PokemonListResponse.PokemonListItem
 import com.sdhong.pokemonapp.model.Pokemons
 import com.sdhong.pokemonapp.remote.RetrofitClient
 import com.sdhong.pokemonapp.remote.api.PokemonApi
+import com.sdhong.pokemonapp.remote.model.PokemonListResponse.PokemonListItem
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.async
+import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
@@ -20,17 +23,22 @@ class AllPokemonViewModel : ViewModel() {
 
     fun getAllPokemon() = viewModelScope.launch {
         val result: List<PokemonListItem> = retrofit.getAllPokemon().results
+        val imgUrlsDeferred = result.map { item -> getImgUrl(item) }
+        val imgUrls = imgUrlsDeferred.awaitAll()
+
         _allPokemon.value = result.mapIndexed { index, item ->
             Pokemon.Normal(
-                id = index,
+                id = index + 1,
                 name = item.name,
-                imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${
-                    item.url.split(
-                        "/"
-                    ).toTypedArray()[6]
-                }.png"
+                imgUrl = imgUrls[index]
             )
         }
+    }
+
+    private fun getImgUrl(item: PokemonListItem): Deferred<String> = viewModelScope.async {
+        val id = item.url.split("/")[6].toInt()
+        val imgUrl = retrofit.getPokemonDetail(id).sprites.imgUrl
+        return@async imgUrl
     }
 
     fun onPokemonClick(
